requirements.txt
numpy==1.26.4
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
Pillow==10.3.0